fix: get react file-upload example working with vite
`create-react-app` projects don't have the proper loaders
to handle loading typescript in dependencies. `vite` is used in our other examples so it makes sense to me that we'd use it for react as well.

putting this in a separate PR before I update all the other examples to get feedback before committing to this approach, and to avoid too many merge conflicts with the efforts underway to update examples by @cmunns and @natevw