fix: Fix docs link from react-uploads-list to keyring page (#196)
(this PR replaces https://github.com/web3-storage/w3ui/pull/151 which
was against a now-obsolete branch)


At least in certain contexts (specifically I was viewing from the URL
<https://github.com/web3-storage/w3ui/blob/w3ui/docs/react-uploads-list.md>
via the `next` branch) the useKeyring documentation link led to a 404
when clicked. Adding the `.md` extension seems to be all that's needed.