chore: use `prepare` script to auto-build after install
Moves the `build` script to `prepare` which is the lifecycle hook that automatically runs after `install`.

This gives us one fewer steps to get started and is a good habit as prepare also runs if folks depend on your modules from a git branch or similar.

Also updates an old ref to `npm` in the scripts and adds us as maintainers!

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>