feat: get a version of UploadsListRoot working
I was not able to get the render props functionality of ariakit-react-utils
working as expected so I did the types manually and dropped support
for the `as` property for this particular case - UploadsListRoot will
always be rendered as a Fragment.

I think this is fine because users can always just add markup inside the render prop function.