feat: support `as` prop in uploader component (#236)
follow the pattern set by HeadlessUI and others of supporting the very
useful "as" property in the uploader component. this:

1. lets users override default tag types (eg, changing a `button` to an
`a` tag)
2. lets users provide custom components to be used as the root element
of our headless components
3. lets users elide a root tag entirely by passing `Fragment` to `as`

See
https://headlessui.com/react/menu#rendering-a-different-element-for-a-component
for more documentation on how this is commonly used

This introduces a new dependency on ariakit-react-utils
(https://ariakit.org/) to avoid reproducing the intricate puzzle of
types in the `createComponent` and `createElement` functions from that
library. I suspect other utilities will be useful as we improve the
accessibility defaults of this library so I think this dependency is
worth adding, but we could copy/paste the specific code we need with
proper attribution if we want to avoid pulling in the whole library.

resolves #235

Co-authored-by: Alan Shaw <alan.shaw@protocol.ai>
Co-authored-by: Yusef Napora <yusef@napora.org>
Co-authored-by: Nathan Vander Wilt <natevw@yahoo.com>