feat: add story for SpaceList
take a slightly different approach from the other stories - wrap
SpaceList in a component specifically made for StoryBook testing
that passes values to a Context Provider and renders the component
under test inside the Provider.

This lets us craft an interface in StoryBook that will make sense
to designers and other folks.