wip: set page size in hook
this is broken as implemented, for a couple reasons:

a) in the best case scenario, there's a flicker in the list as it displays the default page (with more than 10 items in it) for a second before the 10-item page loads and displays
b) in the worst case, the 10-item page request actually completes first!

I suspect the abort controller is not working quite the way we expect, but I'm also not sure that fixing it would solve all these problems.

for now, will probably just revert to setting page size staticly in the provider