feat: allow users to set page size in W3APIProvider
Allow users to set page size in W3APIProvider.

It feels like it would be cleaner to make it possible to set page size
in the `useUploadsList` hook, but a bit of work on that
revealed some complexity that will take a bit more planning
and design to get right - see https://github.com/web3-storage/w3ui/commit/e50e02594bdf608b9add0241f22bed09c7dc2902
for the work I did toward this and a short description of the issues
I ran into.

For now I think this is good enough to ship, and I'd like to think more carefully about how we handle state in the UploadsList in future versions of these libs.