feat: implement reverse paging
Use the new `pre` caveat and `startCursor` and `endCursor` from https://github.com/web3-storage/w3infra/pull/139 to implement reverse paging.

One unsatisfying issue with this is that paging backwards reverses the
order of items in the uploads list, and that edge conditions
behave fairly confusingly.

Not entirely sure what to do about this yet, so pushing up for some feedback.

This currently only works with a bunch of custom service config and `file:/`
dependencies that I am not pushing up for now.