chore: remove prettier, estlint --fix
prettier and eslint were two wolves fighting in this repository.
`eslint --fix` does a substantial amount of what prettier is designed
to do, and the two projects seem to have some major disagreements about
proper TS formatting and it's easier to just have one
tool in the mix. standardize on eslint and make the code conform

also get eslint working in all of the package repos, huzzah!