chore: Move all styling to idiomatic TailwindCSS.
Since we're no longer using the external w3ui React "customizable" component
library, we can use idiomatic TailwindCSS across the board.

A few years ago I would have considered this a step backwards, moving from semantic classes
to tailwind utility classes sprinkled everywhere. While there
may be some opportunities for abstraction in a few places, Tailwind's docs make the case
for being extremely conservative with this sort of CSS-level componentization:

https://tailwindcss.com/docs/reusing-styles

I have created a `w3ui-button` class, but it probably needs to be pared
down - evidence for this can be found in the authenticator's "Register" button
in which I needed to copy most of the styles from the `w3ui-button` class
to customize it the way I wanted.