chore: eject @w3ui/react from w3console (#383)
We don't have the bandwidth to support @w3ui/react anymore, so migrate
w3console to use the "headless" components directly.

Importantly, this means we can use idiomatic TailwindCSS across the
board rather than the hybrid raw-CSS/Tailwind approach we had been
using.

A few years ago I would have considered the CSS changes here a step
backwards, moving from semantic classes to Tailwind utility classes
sprinkled everywhere. While there may be some opportunities for
CSS-class-based abstraction in a few places, Tailwind's docs make the
case for being extremely conservative with this type of
componentization:

https://tailwindcss.com/docs/reusing-styles

I have created a `w3ui-button` class, but it probably needs to be pared
down - evidence for this can be found in the authenticator's "Register"
button in which I needed to copy most of the styles from the
`w3ui-button` class to customize it the way I wanted. I also created an
`authenticator` class, but it may be a better idea to create a React
component for this.

I have this configured to merge to `feat/pre-paging` for now to make
this PR cleaner, but won't merge it until that hits `main` and this gets
automatically retargetted.