feat: add terms of service page
This required adding react-router-dom so we can have more than one page.

This definitely got me wondering if we shouldn't just move over to Next or Remix or someting that handles this and a lot more, but that probably requires a larger discussion. We should also discuss the middle ground of https://vite-plugin-ssr.com/ while we're at it.