feat: add support for getting an account's plan
add a new function to the keyring context that lets clients get an account's billing plan

After chatting with @gozala I'm removing support for umd and cjs builds because the fr32-sha2-256-trunc254-padded-binary-tree-multihash package has a top-level await and they are incompatible with that. Our feeling is that supporting CJS or UMD users of w3ui is not a high priority, but curious how the rest of the team feels!