feat: add a logout function  (#595)
This resets the store, deleting the local agent's private keys and saved
delegations. This is better than the previous implementation, and is
appropriate for use cases like using console on a public or shared
computer.

This PR also has a small tweak to replace JSX.Element with ReactNode,
which I've found works better in downstream applications.