fix: add missing receiptsEndpoint to client instantiation (#647)
## Description

This PR adds the `receiptsEndpoint` to the `w3up-client` instantiation.

## Motivation

Uploading in staging is currently broken. The console shows an error
when attempting to retrieve the receipt.

The receipts endpoint being used for staging is incorrect due to a
missing configuration. This causes the value to default to a hardcoded
production endpoint.